Fix some Composite Compliance problems with binary_cross_entropy backward
This PR fixes composite compliance problems with:
- binary_cross_entropy's backward formula
- binary_cross_entropy_with_logits's backward formula
- binary_cross_entropy's double backward formula

It does so by adding checks for areAnyTensorSubclassLike.

Test Plan:
- I tested everything with functorch.
- We are going to do https://github.com/pytorch/pytorch/issues/69530 in
the future so we have a way of testing this in core. I need the
binary_cross_entropy ones for something right now and didn't want to
wait until we come up with a solution for #69530.

ghstack-source-id: 5ab1dcba8b863266f8463a25ed357f660ae4cddc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/70198